Layout_TestCase/fast/xmlhttprequest/xmlhttprequest-responsetype-json-utf16-expected.txt

CONSOLE MESSAGE: line 0: XMLHttpRequest.responseType "json" is not supported.
Tests XMLHttpRequest.responseType of "json" fails to parse if the payload was encoded with UTF-16.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL jsonXHR.response should be null (of type object). Was {
    "a": 123
}
 (of type string).
PASS successfullyParsed is true

TEST COMPLETE
